--- db.json
@@ -1,21 +1,0 @@
-{
-  "users": [
-    {
-      "id": 1,
-      "invite_code": 11,
-      "email":"admin@gmile.com",
-      "password":"admin"
-      
-    }
-  ],
-  "comments": [
-    {
-      "id": 1,
-      "body": "some comment",
-      "postId": 1
-    }
-  ],
-  "profile": {
-    "name": "typicode"
-  }
-}
