+++ src/App.js
@@ -1,21 +1,14 @@
 import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
 import './App.css';
-import 'materialize-css';
-import Authenticate from './components/context/Authenticate';
+import FirstPage from './components/FirstPage'
 
-import Router from './components/routes/Router';
 
 function App() {
-	return (
-		<div className="App">
-			<BrowserRouter>
-				<Authenticate>
-					<Router />
-				</Authenticate>
-			</BrowserRouter>
-		</div>
-	);
+  return (
+    <div className="App">
+      <FirstPage/>
+    </div>
+  );
 }
 
 export default App;
