--- src/components/Invoice.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function() {
-	return (
-		<div>
-			<p>I am invoice</p>
-		</div>
-	);
-}
