--- src/components/Packages.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Packages() {
-    return (
-        <div>
-            <p>These are the packages</p>
-        </div>
-    )
-}
