--- src/components/SideBar.jsx
@@ -1,34 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-const SideBar=() =>{
-    return (
-        <div className="sideBar">
-				<div>
-					<div className="header">
-						<i className="fas fa-user-tie fa-3x" /><div><p>Kaliisa<br></br>Admin</p></div>
-					</div>
-				</div>
-				<div className="menu">
-					<div className="menuTitle">
-						<i className="fas fa-home" /><Link to="dashboard">Dash Board</Link>
-					</div>
-					<div className="menuList">
-						<div>
-							<Link to="/suppliers">Suppliers</Link>
-						</div>
-						<div>
-							<Link to="/hub-managers">Hub Managers</Link>
-						</div>
-						<div>
-							<Link to="/invoices">Invoices</Link>
-						</div>
-						<div>
-							<Link to="loaders">Loaders</Link>
-						</div>
-					</div>
-				</div>
-			</div>
-    )
-}
-export default SideBar
