+++ src/components/SignOut.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function SignOut() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
