+++ src/components/SignUp.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export default function Dashboard() {
+export default function SignUp() {
     return (
         <div>
