+++ src/components/Supplier.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 export default function Supplier() {
     return (
         <div>
-            <p>I am a supplier</p>
+            
         </div>
     )
 }
